# backend/docs/swagger.yaml
basePath: /
definitions:
...
  domain.AIFeedbackSettings:
    properties:
      ai_feedback_type:
        items:
          type: string
        type: array
      is_enabled:
        type: boolean
    type: object
...
  domain.AppDetailResp:
    properties:
      id:
        type: string
      kb_id:
        type: string
      name:
        type: string
      recommend_nodes:
        items:
          $ref: '#/definitions/domain.RecommendNodeListResp'
        type: array
      settings:
        $ref: '#/definitions/domain.AppSettingsResp'
      type:
        $ref: '#/definitions/domain.AppType'
    type: object
...
  domain.AppSettings:
    properties:
      ai_feedback_settings:
        allOf:
        - $ref: '#/definitions/domain.AIFeedbackSettings'
        description: AI feedback
      body_code:
        type: string
      btns:
        items: {}
        type: array
      catalog_settings:
        allOf:
        - $ref: '#/definitions/domain.CatalogSettings'
        description: catalog settings
      contribute_settings:
        $ref: '#/definitions/domain.ContributeSettings'
      conversation_setting:
        $ref: '#/definitions/domain.ConversationSetting'
      copy_setting:
        allOf:
        - $ref: '#/definitions/consts.CopySetting'
        enum:
        - ""
        - append
        - disabled
      desc:
        description: seo
        type: string
      dingtalk_bot_client_id:
        type: string
      dingtalk_bot_client_secret:
        type: string
      dingtalk_bot_is_enabled:
        description: DingTalkBot
        type: boolean
      dingtalk_bot_template_id:
        type: string
      disclaimer_settings:
        allOf:
        - $ref: '#/definitions/domain.DisclaimerSettings'
        description: Disclaimer Settings
      discord_bot_is_enabled:
        description: DisCordBot
        type: boolean
      discord_bot_token:
        type: string
      document_feedback_is_enabled:
        description: document feedback
        type: boolean
      feishu_bot_app_id:
        type: string
      feishu_bot_app_secret:
        type: string
      feishu_bot_is_enabled:
        description: FeishuBot
        type: boolean
      footer_settings:
        allOf:
        - $ref: '#/definitions/domain.FooterSettings'
        description: footer settings
      head_code:
        description: inject code
        type: string
      home_page_setting:
        $ref: '#/definitions/consts.HomePageSetting'
      icon:
        type: string
      keyword:
        type: string
      lark_bot_settings:
        allOf:
        - $ref: '#/definitions/domain.LarkBotSettings'
        description: LarkBot
      mcp_server_settings:
        allOf:
        - $ref: '#/definitions/domain.MCPServerSettings'
        description: MCP Server Settings
      openai_api_bot_settings:
        allOf:
        - $ref: '#/definitions/domain.OpenAIAPIBotSettings'
        description: OpenAI API Bot settings
      recommend_node_ids:
        items:
          type: string
        type: array
      recommend_questions:
        items:
          type: string
        type: array
      search_placeholder:
        type: string
      stats_setting:
        $ref: '#/definitions/domain.StatsSetting'
      theme_and_style:
        $ref: '#/definitions/domain.ThemeAndStyle'
      theme_mode:
        description: theme
        type: string
      title:
        description: nav
        type: string
      watermark_content:
        type: string
      watermark_setting:
        allOf:
        - $ref: '#/definitions/consts.WatermarkSetting'
        enum:
        - ""
        - hidden
        - visible
      web_app_comment_settings:
        allOf:
        - $ref: '#/definitions/domain.WebAppCommentSettings'
        description: webapp comment settings
      web_app_custom_style:
        allOf:
        - $ref: '#/definitions/domain.WebAppCustomSettings'
        description: WebAppCustomStyle
      web_app_landing_configs:
        description: WebAppLandingConfigs
        items:
          $ref: '#/definitions/domain.WebAppLandingConfig'
        type: array
      web_app_landing_theme:
        $ref: '#/definitions/domain.WebAppLandingTheme'
      wechat_app_advanced_setting:
        $ref: '#/definitions/domain.WeChatAppAdvancedSetting'
      wechat_app_agent_id:
        type: string
      wechat_app_corpid:
        type: string
      wechat_app_encodingaeskey:
        type: string
      wechat_app_is_enabled:
        description: WechatAppBot 企业微信机器人
        type: boolean
      wechat_app_secret:
        type: string
      wechat_app_token:
        type: string
      wechat_official_account_app_id:
        type: string
      wechat_official_account_app_secret:
        type: string
      wechat_official_account_encodingaeskey:
        type: string
      wechat_official_account_is_enabled:
        description: WechatOfficialAccount
        type: boolean
      wechat_official_account_token:
        type: string
      wechat_service_contain_keywords:
        items:
          type: string
        type: array
      wechat_service_corpid:
        type: string
      wechat_service_encodingaeskey:
        type: string
      wechat_service_equal_keywords:
        items:
          type: string
        type: array
      wechat_service_is_enabled:
        description: WechatServiceBot
        type: boolean
      wechat_service_logo:
        type: string
      wechat_service_secret:
        type: string
      wechat_service_token:
        type: string
      wecom_ai_bot_settings:
        allOf:
        - $ref: '#/definitions/domain.WecomAIBotSettings'
        description: WecomAIBotSettings 企业微信智能机器人
      welcome_str:
        description: welcome
        type: string
      widget_bot_settings:
        allOf:
        - $ref: '#/definitions/domain.WidgetBotSettings'
        description: Widget bot settings
    type: object
  domain.AppSettingsResp:
    properties:
      ai_feedback_settings:
        allOf:
        - $ref: '#/definitions/domain.AIFeedbackSettings'
        description: AI feedback
      body_code:
        type: string
      btns:
        items: {}
        type: array
      catalog_settings:
        allOf:
        - $ref: '#/definitions/domain.CatalogSettings'
        description: catalog settings
      contribute_settings:
        $ref: '#/definitions/domain.ContributeSettings'
      conversation_setting:
        $ref: '#/definitions/domain.ConversationSetting'
      copy_setting:
        $ref: '#/definitions/consts.CopySetting'
      desc:
        description: seo
        type: string
      dingtalk_bot_client_id:
        type: string
      dingtalk_bot_client_secret:
        type: string
      dingtalk_bot_is_enabled:
        description: DingTalkBot
        type: boolean
      dingtalk_bot_template_id:
        type: string
      disclaimer_settings:
        allOf:
        - $ref: '#/definitions/domain.DisclaimerSettings'
        description: Disclaimer Settings
      discord_bot_is_enabled:
        description: DisCordBot
        type: boolean
      discord_bot_token:
        type: string
      document_feedback_is_enabled:
        description: document feedback
        type: boolean
      feishu_bot_app_id:
        type: string
      feishu_bot_app_secret:
        type: string
      feishu_bot_is_enabled:
        description: FeishuBot
        type: boolean
      footer_settings:
        allOf:
        - $ref: '#/definitions/domain.FooterSettings'
        description: footer settings
      head_code:
        description: inject code
        type: string
      home_page_setting:
        $ref: '#/definitions/consts.HomePageSetting'
      icon:
        type: string
      keyword:
        type: string
      lark_bot_settings:
        allOf:
        - $ref: '#/definitions/domain.LarkBotSettings'
        description: LarkBot
      mcp_server_settings:
        allOf:
        - $ref: '#/definitions/domain.MCPServerSettings'
        description: MCP Server Settings
      openai_api_bot_settings:
        allOf:
        - $ref: '#/definitions/domain.OpenAIAPIBotSettings'
        description: OpenAI API settings
      recommend_node_ids:
        items:
          type: string
        type: array
      recommend_questions:
        items:
          type: string
        type: array
      search_placeholder:
        type: string
      stats_setting:
        $ref: '#/definitions/domain.StatsSetting'
      theme_and_style:
        $ref: '#/definitions/domain.ThemeAndStyle'
      theme_mode:
        description: theme
        type: string
      title:
        description: nav
        type: string
      watermark_content:
        type: string
      watermark_setting:
        $ref: '#/definitions/consts.WatermarkSetting'
      web_app_comment_settings:
        allOf:
        - $ref: '#/definitions/domain.WebAppCommentSettings'
        description: webapp comment settings
      web_app_custom_style:
        allOf:
        - $ref: '#/definitions/domain.WebAppCustomSettings'
        description: WebAppCustomStyle
      web_app_landing_configs:
        description: WebApp Landing Settings
        items:
          $ref: '#/definitions/domain.WebAppLandingConfigResp'
        type: array
      web_app_landing_theme:
        $ref: '#/definitions/domain.WebAppLandingTheme'
      wechat_app_advanced_setting:
        $ref: '#/definitions/domain.WeChatAppAdvancedSetting'
      wechat_app_agent_id:
        type: string
      wechat_app_corpid:
        type: string
      wechat_app_encodingaeskey:
        type: string
      wechat_app_is_enabled:
        description: WechatAppBot
        type: boolean
      wechat_app_secret:
        type: string
      wechat_app_token:
        type: string
      wechat_official_account_app_id:
        type: string
      wechat_official_account_app_secret:
        type: string
      wechat_official_account_encodingaeskey:
        type: string
      wechat_official_account_is_enabled:
        description: WechatOfficialAccount
        type: boolean
      wechat_official_account_token:
        type: string
      wechat_service_contain_keywords:
        items:
          type: string
        type: array
      wechat_service_corpid:
        type: string
      wechat_service_encodingaeskey:
        type: string
      wechat_service_equal_keywords:
        items:
          type: string
        type: array
      wechat_service_is_enabled:
        description: WechatServiceBot
        type: boolean
      wechat_service_logo:
        type: string
      wechat_service_secret:
        type: string
      wechat_service_token:
        type: string
      wecom_ai_bot_settings:
        $ref: '#/definitions/domain.WecomAIBotSettings'
      welcome_str:
        description: welcome
        type: string
      widget_bot_settings:
        allOf:
        - $ref: '#/definitions/domain.WidgetBotSettings'
        description: WidgetBot
    type: object
  domain.AppType:
    enum:
    - 1
    - 2
    - 3
    - 4
    - 5
    - 6
    - 7
    - 8
    - 9
    - 10
    - 11
    - 12
    format: int32
    type: integer
    x-enum-varnames:
    - AppTypeWeb
    - AppTypeWidget
    - AppTypeDingTalkBot
    - AppTypeFeishuBot
    - AppTypeWechatBot
    - AppTypeWechatServiceBot
    - AppTypeDisCordBot
    - AppTypeWechatOfficialAccount
    - AppTypeOpenAIAPI
    - AppTypeWecomAIBot
    - AppTypeLarkBot
    - AppTypeMcpServer
...
  domain.CatalogSettings:
    properties:
      catalog_folder:
        description: '1: 展开, 2: 折叠, default: 1'
        type: integer
      catalog_visible:
        description: '1: 显示, 2: 隐藏, default: 1'
        type: integer
      catalog_width:
        description: '200 - 300, default: 260'
        type: integer
    type: object
...
  domain.ContributeSettings:
    properties:
      is_enable:
        type: boolean
    type: object
...
  domain.ConversationSetting:
    properties:
      copyright_hide_enabled:
        type: boolean
      copyright_info:
        type: string
    type: object
...
  domain.DisclaimerSettings:
    properties:
      content:
        type: string
    type: object
...
  domain.OpenAIAPIBotSettings:
    properties:
      is_enabled:
        type: boolean
      secret_key:
        type: string
    type: object
...
  domain.PWResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  domain.RecommendNodeListResp:
    properties:
      emoji:
        type: string
      id:
        type: string
      name:
        type: string
      nav_id:
        type: string
      nav_name:
        type: string
      parent_id:
        type: string
      permissions:
        $ref: '#/definitions/domain.NodePermissions'
      position:
        type: number
      recommend_nodes:
        items:
          $ref: '#/definitions/domain.RecommendNodeListResp'
        type: array
      summary:
        type: string
      type:
        $ref: '#/definitions/domain.NodeType'
    type: object
  domain.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  domain.UpdateAppReq:
    properties:
      kb_id:
        type: string
      name:
        type: string
      settings:
        $ref: '#/definitions/domain.AppSettings'
    type: object
...
info:
  contact: {}
  description: panda-wiki API documentation
  title: panda-wiki API
  version: "1.0"
paths:
  /api/v1/app:
    delete:
      consumes:
      - application/json
      description: Delete app
      parameters:
      - description: kb id
        in: query
        name: kb_id
        required: true
        type: string
      - description: app id
        in: query
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - bearerAuth: []
      summary: Delete app
      tags:
      - app
    put:
      consumes:
      - application/json
      description: Update app
      parameters:
      - description: id
        in: query
        name: id
        required: true
        type: string
      - description: app
        in: body
        name: app
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateAppReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      security:
      - bearerAuth: []
      summary: Update app
      tags:
      - app
  /api/v1/app/detail:
    get:
      consumes:
      - application/json
      description: Get app detail
      parameters:
      - description: kb id
        in: query
        name: kb_id
        required: true
        type: string
      - description: app type
        in: query
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.PWResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.AppDetailResp'
              type: object
      security:
      - bearerAuth: []
      summary: Get app detail
      tags:
      - app
...
  /share/v1/app/wechat/service/answer:
    get:
      consumes:
      - application/json
      description: GetWechatAnswer
      parameters:
      - description: conversation id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      summary: GetWechatAnswer
      tags:
      - Wechat
...
securityDefinitions:
  bearerAuth:
    description: Type "Bearer" + a space + your token to authorize
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"